+++ src/manager.rs
@@ -2,12 +2,16 @@ use lazy_static::lazy_static;
 use nng::*;
 use serde::{Deserialize, Serialize};
 use std::collections::HashSet;
+use std::io::BufWriter;
 use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
 use std::sync::Arc;
 use std::{thread, time};
 
 use crate::goose::GooseRequest;
-use crate::metrics::{self, GooseRequestMetrics, GooseTaskMetric, GooseTaskMetrics};
+use crate::metrics::{
+    self, GooseErrorMetric, GooseErrorMetrics, GooseRequestMetrics, GooseTaskMetric,
+    GooseTaskMetrics,
+};
 use crate::util;
 use crate::worker::GaggleMetrics;
 use crate::{GooseAttack, GooseConfiguration, GooseUserCommand};
@@ -147,6 +151,19 @@ fn merge_requests_from_worker(
     merged_request
 }
 
+/// Merge per-Worker errors into global Manager metrics
+fn merge_errors_from_worker(
+    manager_error: &GooseErrorMetric,
+    worker_error: &GooseErrorMetric,
+) -> GooseErrorMetric {
+    // Make a mutable copy where we can merge things
+    let mut merged_error = manager_error.clone();
+    // Add in how many additional times this happened on the Worker.
+    merged_error.occurrences += worker_error.occurrences;
+    // Nothing else changes, so return the merged error.
+    merged_error
+}
+
 /// Helper to send EXIT command to worker.
 fn tell_worker_to_exit(server: &Socket) -> bool {
     let mut message = Message::new();
@@ -213,6 +230,27 @@ fn merge_task_metrics(goose_attack: &mut GooseAttack, tasks: GooseTaskMetrics) {
     }
 }
 
+/// Helper to merge in errors from the Worker.
+fn merge_error_metrics(goose_attack: &mut GooseAttack, errors: GooseErrorMetrics) {
+    if !errors.is_empty() {
+        debug!("errors received: {:?}", errors.len());
+        for (error_key, error) in errors {
+            trace!("error_key: {}", error_key);
+            let merged_error;
+            if let Some(parent_error) = goose_attack.metrics.errors.get(&error_key) {
+                merged_error = merge_errors_from_worker(parent_error, &error);
+            } else {
+                // First time seeing this error, simply insert it.
+                merged_error = error.clone();
+            }
+            goose_attack
+                .metrics
+                .errors
+                .insert(error_key.to_string(), merged_error);
+        }
+    }
+}
+
 /// Main manager loop.
 pub async fn manager_main(mut goose_attack: GooseAttack) -> GooseAttack {
     // Creates a TCP address.
@@ -220,7 +258,7 @@ pub async fn manager_main(mut goose_attack: GooseAttack) -> GooseAttack {
         "tcp://{}:{}",
         goose_attack.configuration.manager_bind_host, goose_attack.configuration.manager_bind_port
     );
-    info!("worker connecting to manager at {}", &address);
+    debug!("preparing to listen for workers at: {}", &address);
 
     // Create a Rep0 reply socket.
     let server = Socket::new(Protocol::Rep0)
@@ -250,7 +288,7 @@ pub async fn manager_main(mut goose_attack: GooseAttack) -> GooseAttack {
     let (users_per_worker, mut users_remainder) = distribute_users(&goose_attack);
 
     // A mutable bucket of users to be assigned to workers.
-    let mut available_users = goose_attack.weighted_users.clone();
+    let mut available_users = goose_attack.weighted_gaggle_users.clone();
 
     // Track how many workers we've seen.
     let mut workers: HashSet<Pipe> = HashSet::new();
@@ -419,6 +457,7 @@ pub async fn manager_main(mut goose_attack: GooseAttack) -> GooseAttack {
                         let mut users = Vec::new();
 
                         // Pop users from available_users vector and build worker initializer.
+                        debug!("sending {} users to worker", user_batch);
                         for _ in 1..=user_batch {
                             let user = match available_users.pop() {
                                 Some(u) => u,
@@ -438,14 +477,21 @@ pub async fn manager_main(mut goose_attack: GooseAttack) -> GooseAttack {
                             });
                         }
 
-                        // Send vector of user initializers to worker.
-                        let mut message = Message::new();
+                        // Prepare to serialize the list of users to send to the Worker.
+                        let mut message = BufWriter::new(Message::new());
+
+                        info!("serializing users with serde_cbor...");
                         serde_cbor::to_writer(&mut message, &users)
                             .map_err(|error| eprintln!("{:?}", error))
                             .expect("failed to serialize user initializers");
 
                         info!("sending {} users to worker {}", users.len(), workers.len());
-                        if !send_message_to_worker(&server, message) {
+                        if !send_message_to_worker(
+                            &server,
+                            message
+                                .into_inner()
+                                .expect("failed to extract nng message from buffer"),
+                        ) {
                             // All workers have exited, shut down the load
                             // test.
                             break;
@@ -496,6 +542,10 @@ pub async fn manager_main(mut goose_attack: GooseAttack) -> GooseAttack {
                             GaggleMetrics::Tasks(tasks) => {
                                 merge_task_metrics(&mut goose_attack, tasks)
                             }
+                            // Merge in error metrics from Worker.
+                            GaggleMetrics::Errors(errors) => {
+                                merge_error_metrics(&mut goose_attack, errors)
+                            }
                             // Ignore Worker heartbeats.
                             GaggleMetrics::WorkerInit(_) => (),
                         }
@@ -527,8 +577,8 @@ pub async fn manager_main(mut goose_attack: GooseAttack) -> GooseAttack {
                         break;
                     }
                     if !load_test_finished {
-                        // Sleep half a second then return to the loop.
-                        thread::sleep(time::Duration::from_millis(500));
+                        // Sleep a tenth of a second then return to the loop.
+                        thread::sleep(time::Duration::from_millis(100));
                     }
                 } else {
                     panic!("error receiving user message: {}", e);